Enable releasing as well gtk-common-themes with latest communitheme
We don't rebuild gtk-common-themes as we only want to rebuild it with latest
communitheme snap content that we extract.
Always download the preambule of gtk-common-themes and then, download
from the edge channel its content, stripping communitheme one.
Then, ensure we download communitheme from the current channel and push that
to the store.