Fixed misalignment for small title button in chromium/chrome
This issue appeared on chromium/chrome only, therefore specific code as been
written in _apps.scss. Chrome is a widely used application, therefore I consider
this a valid exception to our idea of not-app-specific changes.
Moreover, the change is basically an implementation of "small/smaller titlebutton"
that might turn out useful for other applications in the future.

closes #412