Made notebook tabs more similar to gnome-terminal's ones
In dark variant, notebook tabs are similar to gnome-terminal's tabs, but
they are style differently. Since gnome-terminal's tabs seems to be
appreciated, this commit style generic notebook tabs in a similar way.

NOTE: generic notebook tabs do not get the orange underline
gnome-terminal has. In fact, while the orange looks good in the
particular usage of gnome-terminal, in the wider variety of usage of
generic notebook tabs, this is not always true and instead orange make
those tabs look heavier.