notebook: refined unchecked tabs Nautilus + Gedit
In applications with a sidebar and status-bar (e.g. Nautilus and Gedit),
having unchecked notebok tabs with the same color as headerbar creates
a unpleasant effect of disconnection between the notebook and the
sidebar/status-bar.

To fix that this commit
- changes the sidebar color in Gedit to bg_color (Nautilus already uses it)
- changes the color for unchecked tabs in Gedit and Nautilus to bg_color
  (the one used by the sidebar and status bar)

Note: general notebook tabs, like the ones in gnome-terminal, keep the
previous headerbar_bg_color

closes #786