Set top-panel always full opaque
Change taken from #1207

Gnome-shell 3.31.90 removed top-bar translucency feature in a way that the
settings we use for the .transparent class are used when the windows are
maximized.

This commit sets the transparency for top-bar at zero always. Since the comeback
of this feature is currently discussed, it is better to keep al the logic in place
so that it will be easy to reuse it later.

closes #1201