Remove bordercolordifference in darkthemeswitches
When the border color of the top border is different than the rest of the border, there are dark and washed out edges in the border top where a darker green shines through the lighter green. Probably because gtk lets the background color shine through the green. which is no problem in the light theme but in the headerbar and the darktheme. We changed this already for the headerbar some months ago, this commit does the same for the dark theme in general.