Fix slider fine-tune
Fine-tune feature currently is not working properly. Instead of making
the slider thicker, it actually changes only the space around causing the
window to shrink.

The fix consists in changing the margin-(top/bottom/left/right) according
to the position of the marks (above/below/etc.).

Note: tested only on slider with marks below for the lack of apps with
the other kind of sliders.