Ensure shell-theme link can be created when using DESTDIR
When creating a package (such as on Arch Linux) using:

`DESTDIR="${pkgdir}" ninja -C build install`

to install all files to `${pkgdir}` for later installation onto an
actual system, ensure the symlink to the shell theme can be created.

The link target should point to a location in `MESON_INSTALL_PREFIX`,
but the link itself should be in `MESON_INSTALL_DESTDIR_PREFIX`.
Furthermore, the link's parent directory needs to be created within
`MESON_INSTALL_DESTDIR_PREFIX` if it does not exist.

Fixes #1583