tweaks: notebook tab hover text color
On hover, notebook tab's text is grey and so it might be mistaken for
disabled.

This change sets for tab's text on hover the same color that unchecked
tab's text have.
Moreover it sets a slightly brighter background color, still less than
checked tabs, to provide more feedback to the user.

closes #1762