tweaks: notebook tab no hover effect in backdrop
For consistency with the rest of the widgets that have no hover effect
in backdrop, prevent ours notebook's tab to have any hover effect in the
same state.

Note: it worths propose to give a backdrop hover effect to all widgets
upstreams, since all gtk widget are still active and fully functional in
backdrop.