shell: improve scrollbar border in dark mode
Using deep dark border for scrollbar makes the scrollbar look way
smaller than in light variant and also clashes with the current border
of overdrive.

This change aligns the active scrollbar border to the overdrive one,
using a darker aubergine instead that a dark color. This, of course,
only for the dark variant.