tweaks: notebook's tab unchecked font weight
Currently, unchecked tab use a normal font weight. While this looks
lighter to the eye and make it easier to spot the checked tab, it also
has an annoying side effect: when the tab is selected, the whole tab
seems to get bigger (not just the text).

Since the drawback of the light font size is more evident than the
intended effect, this commit changes the font-weight from normal to 500,
which is "just above normal, and not bold", even if at a first sight the
difference from actual bold is limited.