common: use $bg_color for sidebar
Nautilus sidebar color is not different enough to create a separation
from the main windows. The two colors are so similar that one might ask
why not making them the same.

The motivation to revert sidebar to $bg_color is to avoid having yet
another color in the palette and to create some matching with the other
main applications with a sidebar (gnome-settings, gedit and tweaks).

closes #1803