calendar-other-month-day color shall not change if the day has an event. (#2380)
"calendar-other-month-day" with event has a different color than
"calendar-other-month-day" without an event.

"calendar-other-month-day" is a special day, like non-working day, and
should not have a different color, or it would be mistaken for an
in-current-month working day.

Prevent this changing the color of a day with an event only if not a
"calendar-other-month-day".

Closes: #2164