Replace single symbolic file with one file for each symbol
Since working in parallel on the same svg file can cause many
synchronization problems, this change is intended to split the
source-symbolic file into its components.

The copies - optimized - of the contents of source-symbolic is already
in icons/Suru/scalable folder, then this change copies these folders inside
icons/src/scalable.
Finally, this change provides a bash script to optimize the SVG files
using svgo, as the previous ruby script did.

render-symbolic-icons.sh can work on a single file, on all the files
into a group (e.g. actions), or on the full list of groups, using the
flags --file, --group, and --all.
See  render-symbolic-icons.sh --help for more info.

The next step of this change would be to remove the, now unnecessary,
resources like the ruby script and rewrite README.