Add desktop-logoff sound
When the sounds were added, the logoff one wasn't. @madsrh says
this was because it'd be better to just log off immediately, and not
wait for a sound to play.

Session managers can handle this stuff, and play the sound
while logging off in the background. But if it doesn't work, the
sound can always be disabled.