debian/rules: Enable user themes support in packages
We used to install the extensions in places where User Themes extension
would look, and we don't want to break this assumption at this level.

It's just better to install all the themes for now, so it would be
possible for the tweakers to have a mixed dark/light setup, despite the
gnome-control-center settings.